feat: fractional / non-uniform klen for the sieve route
--leaked-bits is now a real number (e.g. 2.5). try_sieve builds a
per-signature nonce-length split (make_klen_list: mix of floor/ceil so
the average matches) and emits a distinct klen per signature, letting the
worker's non-uniform-klen basis handle fractional-average leakage.

Exact for integer L; for a supplied *average* the floor/ceil split is
positional, so fractional L is exact only when per-signature leakage is
known or generated to match (see make_klen_list note).

+++ include/recovery_engine.h
@@ -33,8 +33,10 @@ public:
         // the leak is *supplied* (side-channel model), not statistically detected
         // -- essential for deep leaks (L<=3) where too-few-samples detection fails.
         // Builds an MSB profile directly and, with a pubkey present, routes to the
-        // sieving-with-predicate worker. 0 (default) leaves behaviour unchanged.
-        int msb_leaked_bits = 0
+        // sieving-with-predicate worker. May be fractional (e.g. 2.5 -> a per-
+        // signature mix of the two bracketing bounds). 0 (default) leaves
+        // behaviour unchanged.
+        double msb_leaked_bits = 0.0
     );
 
 private:
